Support fling velocity notifications to the clients of both WebGestureCurve and WebInputHandler.
A WebGestureCurve may have an associated velocity.  This patch provides the
necessary hook for such velocities to be sent to the curve's associated
client. Additionally, plumbing is added to pipe these values from
WebCompositorInputHandlerImpl to its associated WebInputHandlerClient. This is
used with overscroll notifications, as in
https://codereview.chromium.org/13375020/, but could prove useful in such cases
as compositor viewport prediction, etc...

BUG=135975
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/13674008

git-svn-id: svn://svn.chromium.org/blink/trunk@149219 bbb929c8-8fbe-4397-9dbb-9b2b20218538