Fix crash in single process mode on Mac.
The single process mode is used for testing. In that mode, WebKit runs on a thread in the browser process. In that mode, this notification is called by the OS on the main thread in the browser process, and not on the WebKit thread. Avoid the crash by early exiting in the callback.

BUG= 126514

Review URL: https://chromiumcodereview.appspot.com/14503008

git-svn-id: svn://svn.chromium.org/blink/trunk@149231 bbb929c8-8fbe-4397-9dbb-9b2b20218538