First round of cleaning up the Android layout test runner
This moves all methods to interact with Android devices out of the
test port and driver, into a separate AndroidCommands class. I also
updated the unittests for the Chromium Android layout test runner
script to (1) pass again, (2) work with this new abstraction, (3) be
more relevant to what's actually being used and (4) be readable.

This patch also fixes a bug (and adds a test) when executing the
run-webkit-tests script with the "--adb-device" argument.

BUG=232044
R=eseidel@chromium.org, jochen@chromium.org

Review URL: https://codereview.chromium.org/14188013

git-svn-id: svn://svn.chromium.org/blink/trunk@149345 bbb929c8-8fbe-4397-9dbb-9b2b20218538