Clean up RenderLayer paintOrderLists code
r149028 added a ton of code to RenderLayer that was more complex
than it needed to be. We should never use NodeLists from inside the
render tree, and there's also no reason to use a HashSet to store
RenderLayer related nodes as we can never see the same one twice.
The code was also passing around the size of Vectors which is silly,
Vectors have size()! :)

This patch simplifies the code a bunch to match best practice in
the render tree.

This patch also cleans up style guide violations such as missing
whitespace around keywords, an improperly indented switch, and
removes the default: clause from the switch statements that match
all cases as this is preferred since adding a new type for the switch
enum would cause compilation failure.

Another advantage of this patch is that paintOrderLists is the last
user of StaticHashSetNodeList. After this patch we can remove
this class.

Review URL: https://chromiumcodereview.appspot.com/14458008

git-svn-id: svn://svn.chromium.org/blink/trunk@149382 bbb929c8-8fbe-4397-9dbb-9b2b20218538