Tweak UMA stat that measures effect of promoting RenderLayers with a webkit-transition
This tweaks the stat that was introduced by http://crrev.com/13620006 so that
only layers that intersect the viewport are considered for promotion.

The data gathered so far indicates that promoting all layers with a
webkit-transition on opacity or transform (regardless of whether the layer
intersects the viewport) has too large an impact on memory: 5% of samples
had a >33% increase in memory, and 1% of samples had a >150% increase in
memory. This is despite the fact that the we are only lower-bounding the impact
(since we're ignoring the cost of layers that overlap the layers we'd like to
 promote).

The change made here will help us determine if only promoting such layers
that intersect the viewport is a reasonable alternative.

BUG=178119

Review URL: https://chromiumcodereview.appspot.com/14656003

git-svn-id: svn://svn.chromium.org/blink/trunk@149466 bbb929c8-8fbe-4397-9dbb-9b2b20218538