Add parsing for <flex>
This change adds parsing, style application and getComputedStyle
code for <flex>. Due to crbug.com/235629 (<dimension> parsing bug),
we are failing in one case (negative <flex>).

GridLength was extended to support <flex> (as an unsigned) and the
layout code was updated to avoid assuming that GridLength is a Length.

While amending RenderGrid, tweaked the ASSERT to underline that we
shouldn't see 'auto' lengths during layout as we convert them in
GridTrackSize.

BUG=235258
TESTS=fast/css-grid-layout/grid-columns-rows-get-set-multiple.html
      fast/css-grid-layout/grid-columns-rows-get-set.html

Review URL: https://chromiumcodereview.appspot.com/14113052

git-svn-id: svn://svn.chromium.org/blink/trunk@149480 bbb929c8-8fbe-4397-9dbb-9b2b20218538