Allow using both EnabledAtRuntime and EnabledPerContext conditions for the same variable.
Added a new variable called @enabledAtRuntimeAndPerContextAttributes . It contains attributes with both type of flags. When this variable is populated, both conditions are generated by the bindings. This patch does not address multiple conditions of the same type, as I am not aware of such a scenario.

BUG=236824

Review URL: https://chromiumcodereview.appspot.com/14577006

git-svn-id: svn://svn.chromium.org/blink/trunk@149499 bbb929c8-8fbe-4397-9dbb-9b2b20218538