Remove setDeviceScaleFactor and replace with a isHighDPI flag.
The PlatformContextSkia::setDeviceScaleFactor call affects only
the drawing of spelling and grammar markers (now in GraphicsContext).
The choice of high or low-res marker is made depending on whether
the scale factor is 1 or 2 (no other choices are allowed).

This patch changes the setting to a flag. It is a tiny step on
the way to merging and rationalizing PlatformContextSkia and
GraphicsContext.

R=jamesr@chromium.org
BUG=237265

Review URL: https://chromiumcodereview.appspot.com/14637008

git-svn-id: svn://svn.chromium.org/blink/trunk@149554 bbb929c8-8fbe-4397-9dbb-9b2b20218538