Get rid of usages of ScriptWrappable::wrapper().
This is needed for getting rid of Persistent handle copying, as we're going to
remove the copy constructor (see bug for more information).

If we need to pass back a Persistent by reference in the V8 API, we can normally
just pass back a Persistent we hold. ScriptWrappable is problematic, because it
does not hold a Persistent, but instead stores the raw pointer. Added
UnsafePersistent to make the conversion from the raw pointer to a Persistent
explicit.

BUG=236290

Review URL: https://chromiumcodereview.appspot.com/14443007

git-svn-id: svn://svn.chromium.org/blink/trunk@149585 bbb929c8-8fbe-4397-9dbb-9b2b20218538