Rename Text::virtualCreate to cloneWithData
virtualCreate was just a way to create an instance of the same type
of Text node but with a different kind of data, we can use this to
implement cloneNode and get rid of CDATASection::cloneNode. It also
makes sense to rename virtualCreate to cloneWithData to clarify what
it's for.

Review URL: https://chromiumcodereview.appspot.com/14840002

git-svn-id: svn://svn.chromium.org/blink/trunk@149633 bbb929c8-8fbe-4397-9dbb-9b2b20218538