Add an option to run-webkit-tests to ignore tests that are flaky on the bots.
This is speculative for now. In the short-term, I think we probably want to
enable this for the trybots/commit-queue. In the long-term, I'm hoping that
we can remove all the flaky lines from TestExpectations and turn off retrying
failing tests when running locally. That way, people are more likely to identify
when their patch causes a test to fail the first time through, but pass the second.

TBR=dpranke@chromium.org

Review URL: https://codereview.chromium.org/14586002

git-svn-id: svn://svn.chromium.org/blink/trunk@149699 bbb929c8-8fbe-4397-9dbb-9b2b20218538