Move PlatformContextSkia into GraphicsContext and delete
This change moves all of the remaining functionality from PlatformContextSkia to
GraphicsContext, and modifies calling code appropriately. PlatformContextSkia
is removed.

Do not consider this patch to be the final word.

The inlining or non-inlining of several single line methods is arbitrary,
or affected by the way that state enums are defined and the classes/structs
for holding state are declared (PlatformContextSkiaState needs enums
currently in GraphicsContext, so cannot be included in GraphicsContext,
which forces some get/set methods into the cpp file).

There are also methods that directly access the canvas that could be
moved around and rationalized. And the code for maintaining the state
stacks is messy and duplicated (see crbug.com/235470). There also
remain opportunities to change or streamline method argument types.
Finally, there is a lack of clarity on the paintingEnabled/Disabled
behavior (crbug.com/238003).

And almost all of GraphicsContext.h is indented one tab too far right.
Now is a good time to fix that.

Longer term it would also be excellent if we could prevent any code from
accessing the canvas or bitmap directly, the way that various WebKit layers
now do.

R=jamesr@chromium.org, pdr@chromium.org
BUG=235465

Review URL: https://codereview.chromium.org/14560005

git-svn-id: svn://svn.chromium.org/blink/trunk@149909 bbb929c8-8fbe-4397-9dbb-9b2b20218538