Fix rebaselining a directory with a reftest in it.
The bug was in the early return. Should have been a continue.
Also, move the code for handling directory rebaselines up into
AbstractParallelRebaselineCommand so that it can be done in parallel.

Needed to write the files in the unittests because now 
AbstractParallelRebaselineCommand will only try to rebaseline
tests that actually exist.

Review URL: https://chromiumcodereview.appspot.com/15017008

git-svn-id: svn://svn.chromium.org/blink/trunk@149910 bbb929c8-8fbe-4397-9dbb-9b2b20218538