Don't recalc style for the whole subtree during animations
SyntheticStyleChange was used by animations to trigger a style recalc on the
element that is being animated, but it was after FullStyleChange in the enum
which meant we'd Force recalc style on all the descendants of the animating
element as well even though it wasn't needed.

Instead we can put SyntheticStyleChange earlier in the enum and only trigger
a local style change on the animating element and let the normal Node::diff
logic decide if we need to recalc our children.

This should greatly improve performance when animating elements that
contain a large number of descendants.

Review URL: https://chromiumcodereview.appspot.com/14588011

git-svn-id: svn://svn.chromium.org/blink/trunk@150098 bbb929c8-8fbe-4397-9dbb-9b2b20218538