Rewrite time complexity tests
For review; needs testing on bots before committing, else will be flaky.

3 major changes, 3 minor changes:
* Rewrites testing (LayoutTests/resources/magnitude.js)
* Sets parameters on individual tests
* Fix test expectations:
  9 tests were marked as flakey, replaced with 1 slow and 3 failures
Also:
* tweak one test so uses a random number for better accuracy
* workaround to make 1 test not Slow
* 1 minor test HTML fix (make valid HTML 5)

Patch is large b/c of rewriting testing almost completely,
and updating all the tests for the new framework.
(If helpful, could break up by turning off all tests instead,
then individually turning them on.)

For testing, diff isn’t v. useful b/c substantially a rewrite;
a portion of code is retained though.
For individual tests and expectations, the diffs are useful:
for individual tests this shows the parameters (and minor changes),
while for test expectations it shows the bottom line.

For magnitude.js (main file), it's 500 lines, of which 1/4 are comments,
mostly explaining the stats and parameters.

It's structured as:
* main loop (Magnitude.run, Magnitude._runTime),
* utility functions,
* auxiliary computations (mostly Magnitude._OLS for ordinary least squares),
* statistical tests:
  Magnitude._isConstant
  Magnitude._isLogarithmic
  Magnitude._isLinear
  Magnitude._isPolynomial

The basic structure (run test repeatedly, time by measuring iterations in
a given time) is unchanged, but the details of measuring and the tests
themselves have changed substantially.

In the stats sections there are a lot of variables, so that the stages of
computations are clear, so the functions are long but not dense.
The actual test is always just "max deviance <= tolerance".

The meat is in the computations; the actual logic is pretty straightforward,
with the hairiest functions being:
* Magnitude._isLogarithmic
  ...as the log testing is statistically trickier
  (iterative least squares, throwing out extreme points)
* Magnitude._runTime
  ...b/c of various tricks to reduce measurement error,
  notably averaging and retries

Beyond review, needs testing on try bots, else likely will fail,
as parameters have only been tested on my workstation and may be
too strict on bots.

Please feel free to get other opinions and suggest other reviewers!

R=ojan@chromium.org
BUG=232244

Review URL: https://chromiumcodereview.appspot.com/13976010

git-svn-id: svn://svn.chromium.org/blink/trunk@150294 bbb929c8-8fbe-4397-9dbb-9b2b20218538