Implement scoped style tree and scoped style resolver.
A scoped style resolver collects matched scoped rules, including :root.
To solve styles for a given element, we goes up the scoped style tree
from a node whose scoping node is just the given element or the last node
enclosing the given element.

To avoid performance regression, a scoped style resolver for :root and
a flag whether there is only a scoped style resolver for :root or not
is cached in ScopedStyleTree.

BUG=229345

Review URL: https://chromiumcodereview.appspot.com/14137033

git-svn-id: svn://svn.chromium.org/blink/trunk@150398 bbb929c8-8fbe-4397-9dbb-9b2b20218538